fix: move form functions to utils
Movi as funções dentro de FormularioConfiguracao para utils.tsx
